+++ monkey/tests/unit_tests/infection_monkey/master/test_automated_master.py
@@ -10,7 +10,7 @@ INTERVAL = 0.001
 
 
 def test_terminate_without_start():
-    m = AutomatedMaster(None, [], None, MagicMock(), [], MagicMock())
+    m = AutomatedMaster(None, [], None, MagicMock(), [])
 
     # Test that call to terminate does not raise exception
     m.terminate()
@@ -30,7 +30,7 @@ def test_stop_if_cant_get_config_from_island(monkeypatch):
     monkeypatch.setattr(
         "infection_monkey.master.automated_master.CHECK_FOR_TERMINATE_INTERVAL_SEC", INTERVAL
     )
-    m = AutomatedMaster(None, [], None, cc, [], MagicMock())
+    m = AutomatedMaster(None, [], None, cc, [])
     m.start()
 
 
@@ -67,5 +67,5 @@ def test_stop_if_cant_get_stop_signal_from_island(monkeypatch, sleep_and_return_
         "infection_monkey.master.automated_master.CHECK_FOR_TERMINATE_INTERVAL_SEC", INTERVAL
     )
 
-    m = AutomatedMaster(None, [], None, cc, [], MagicMock())
+    m = AutomatedMaster(None, [], None, cc, [])
     m.start()