+++ vulture_allowlist.py
@@ -1,8 +1,6 @@
 from agent_plugins.exploiters.hadoop.plugin import Plugin as HadoopPlugin
 from agent_plugins.exploiters.smb.plugin import Plugin as SMBPlugin
 from agent_plugins.exploiters.snmp.src.snmp_exploit_client import SNMPResult
-from agent_plugins.exploiters.ssh.src.ssh_command_builder import build_ssh_command
-from agent_plugins.exploiters.ssh.src.ssh_options import SSHOptions
 from agent_plugins.exploiters.wmi.plugin import Plugin as WMIPlugin
 from agent_plugins.exploiters.zerologon.src.HostExploiter import HostExploiter
 from flask_security import Security
@@ -21,6 +19,7 @@ from infection_monkey.exploit.zerologon import NetrServerPasswordSet, NetrServer
 from infection_monkey.exploit.zerologon_utils.remote_shell import RemoteShell
 from infection_monkey.network.firewall import FirewallApp, WinAdvFirewall, WinFirewall
 from infection_monkey.utils import commands
+from infection_monkey.utils.decorators import request_cache
 from monkey_island.cc.deployment import Deployment
 from monkey_island.cc.models import IslandMode, Machine
 from monkey_island.cc.repositories import IAgentEventRepository, MongoAgentEventRepository
@@ -161,7 +160,4 @@ commands.build_dropper_script_download_command
 commands.build_download_command_windows_powershell_webclient
 commands.build_download_command_windows_powershell_webrequest
 
-# Remove after #3170 is done
-SSHOptions
-SSHOptions.connect_timeout
-build_ssh_command
+request_cache